restructure bounds
