+++ src/coomsuite/bounds/multi_application.py
@@ -2,15 +2,16 @@
 Clingo application class for solving COOM configuration problems with multi-shot solving
 """
 
-from itertools import count, dropwhile
-from typing import Any, Dict, Iterator, List, Optional, Sequence, Set, Tuple, TypeAlias
+from typing import Any, Dict, List, Optional, Sequence, Set, Tuple, TypeAlias
 
 from clingo import Control
 from clingo.symbol import Function, Number, Symbol, parse_term
 
-from .application import COOMSolverApp
-from .preprocess import preprocess
-from .utils import get_encoding
+from coomsuite.application import COOMSolverApp
+from coomsuite.preprocess import preprocess
+from coomsuite.utils import get_encoding
+
+from . import get_bound_iter, next_bound_converge
 
 ProgPart: TypeAlias = Tuple[str, List[Symbol]]
 
@@ -34,43 +35,6 @@ def _get_fact_name_and_args(fact: str) -> Tuple[str, List[Symbol]]:
     """
     x = parse_term(fact[:-1])
     return (x.name, x.arguments)
-
-
-def _exponential_iter() -> Iterator[int]:
-    n = 0
-    while True:
-        yield 2**n
-        n += 1
-
-
-def get_bound_iter(algorithm: str, start: int) -> Iterator[int]:
-    """
-    Get an iterator over the bounds for a selected algorithm
-
-    Note that the iterator starts after the start value.
-    """
-    iterator: Iterator[int]
-    if algorithm == "linear":
-        iterator = count(start + 1)
-    elif algorithm == "exponential":
-        iterator = dropwhile(lambda x: x <= start, _exponential_iter())
-    else:
-        raise ValueError(f"unknown algorithm for bound iter: {algorithm}")
-
-    return iterator
-
-
-def next_bound_converge(unsat_bound: int, sat_bound: int) -> Optional[int]:
-    """
-    Determine the next bound (between unsat_bound and sat_bound) while converging to the optimal bound
-
-    Returns:
-        Optional[int]: The next bound to check, or None if sat_bound is already the optimal bound
-    """
-    if unsat_bound + 1 == sat_bound:
-        return None
-
-    return (unsat_bound + sat_bound) // 2
 
 
 class COOMMultiSolverApp(COOMSolverApp):
